requirements.txt
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==1.1.8
paho-mqtt==1.5.1
pkg-resources==0.0.0
RPi.GPIO==0.7.0
spidev==3.5